+++ CHANGES.rst
@@ -12,6 +12,7 @@ New features
 - Access Datalabs datafiles without download or copy with option ``--data labs``.
 - Retrieve tile indices with ``dss`` provider (option ``--tiling`` is kept as an optimization).
 - Add option ``-q tiles`` to query just tile indices and not filenames.
+- Add option ``--survey`` to select and sort processing modes.
 
 ``azul process``